Remove SettingsActivity and merge everything into one nav graph
