--- app/src/main/res/layout/activity_settings.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/parent_layout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:theme="@style/SettingsTheme">
-
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/face_fragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/colorPrimaryDark" />
-
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/settings_nav_host_fragment"
-        android:name="androidx.navigation.fragment.NavHostFragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:defaultNavHost="true"
-        app:navGraph="@navigation/settings_nav_graph"/>
-
-    <com.willowtree.vocable.customviews.PointerView
-        android:id="@+id/pointer_view"
-        android:layout_width="@dimen/pointer_view_width_height"
-        android:layout_height="@dimen/pointer_view_width_height"
-        android:background="@drawable/pointer_background"
-        android:elevation="@dimen/settings_pointer_elevation"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <include
-        android:id="@+id/error_view"
-        layout="@layout/error_layout"
-        android:visibility="gone" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
