+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -6,6 +6,102 @@
     app:startDestination="@id/presetsFragment">
 
     <fragment
+        android:id="@+id/editCategoriesFragment"
+        android:name="com.willowtree.vocable.settings.EditCategoriesFragment"
+        android:label="EditCategoriesFragment"
+        tools:layout="@layout/fragment_edit_categories">
+        <action
+            android:id="@+id/action_editCategoriesFragment_to_editCategoriesKeyboardFragment"
+            app:destination="@id/editCategoriesKeyboardFragment" />
+        <action
+            android:id="@+id/action_editCategoriesFragment_to_editCategoryOptionsFragment"
+            app:destination="@id/editCategoryOptionsFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/editCategoriesKeyboardFragment"
+        android:name="com.willowtree.vocable.settings.EditCategoriesKeyboardFragment"
+        android:label="EditCategoriesKeyboardFragment"
+        tools:layout="@layout/fragment_edit_keyboard">
+        <argument
+            android:name="category"
+            android:defaultValue="@null"
+            app:argType="com.willowtree.vocable.room.Category"
+            app:nullable="true" />
+    </fragment>
+    <fragment
+        android:id="@+id/editCategoryOptionsFragment"
+        android:name="com.willowtree.vocable.settings.EditCategoryOptionsFragment"
+        android:label="EditCategoryOptionsFragment">
+        <argument
+            android:name="category"
+            app:argType="com.willowtree.vocable.room.Category" />
+        <action
+            android:id="@+id/action_editCategoryOptionsFragment_to_editCategoriesKeyboardFragment"
+            app:destination="@id/editCategoriesKeyboardFragment" />
+        <action
+            android:id="@+id/action_editCategoryOptionsFragment_to_addPhraseKeyboardFragment"
+            app:destination="@id/addPhraseKeyboardFragment" />
+        <action
+            android:id="@+id/action_editCategoryOptionsFragment_to_editPhrasesKeyboardFragment"
+            app:destination="@id/editPhrasesKeyboardFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/editPresetsFragment"
+        android:name="com.willowtree.vocable.settings.EditPresetsFragment"
+        android:label="EditPresetsFragment"
+        tools:layout="@layout/fragment_edit_presets">
+        <action
+            android:id="@+id/action_editPresetsFragment_to_editPhrasesKeyboardFragment"
+            app:destination="@id/editPhrasesKeyboardFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/sensitivityFragment"
+        android:name="com.willowtree.vocable.settings.SensitivityFragment"
+        android:label="SensitivityFragment"
+        tools:layout="@layout/fragment_timing_sensitivity" />
+    <fragment
+        android:id="@+id/selectionModeFragment"
+        android:name="com.willowtree.vocable.settings.SelectionModeFragment"
+        android:label="SelectionModeFragment"
+        tools:layout="@layout/fragment_selection_mode" />
+    <fragment
+        android:id="@+id/editPhrasesKeyboardFragment"
+        android:name="com.willowtree.vocable.settings.EditPhrasesKeyboardFragment"
+        android:label="EditPhrasesKeyboardFragment"
+        tools:layout="@layout/fragment_edit_phrases">
+        <argument
+            android:name="phrase"
+            android:defaultValue="@null"
+            app:argType="com.willowtree.vocable.room.Phrase"
+            app:nullable="true" />
+    </fragment>
+    <fragment
+        android:id="@+id/addPhraseKeyboardFragment"
+        android:name="com.willowtree.vocable.settings.AddPhraseKeyboardFragment"
+        android:label="AddPhraseKeyboardFragment">
+        <argument
+            android:name="category"
+            app:argType="com.willowtree.vocable.room.Category" />
+    </fragment>
+    <fragment
+        android:id="@+id/settingsFragment"
+        android:name="com.willowtree.vocable.settings.SettingsFragment"
+        android:label="SettingsFragment"
+        tools:layout="@layout/fragment_settings">
+        <action
+            android:id="@+id/action_settingsFragment_to_editPresetsFragment"
+            app:destination="@id/editPresetsFragment" />
+        <action
+            android:id="@+id/action_settingsFragment_to_sensitivityFragment"
+            app:destination="@id/sensitivityFragment" />
+        <action
+            android:id="@+id/action_settingsFragment_to_selectionModeFragment"
+            app:destination="@id/selectionModeFragment" />
+        <action
+            android:id="@+id/action_settingsFragment_to_editCategoriesFragment"
+            app:destination="@id/editCategoriesFragment" />
+    </fragment>
+    <fragment
         android:id="@+id/presetsFragment"
         android:name="com.willowtree.vocable.presets.PresetsFragment"
         android:label="PresetsFragment"
@@ -14,8 +110,8 @@
             android:id="@+id/action_presetsFragment_to_keyboardFragment"
             app:destination="@id/keyboardFragment" />
         <action
-            android:id="@+id/action_presetsFragment_to_settingsActivity"
-            app:destination="@id/settingsActivity" />
+            android:id="@+id/action_presetsFragment_to_settingsFragment"
+            app:destination="@id/settingsFragment" />
     </fragment>
     <fragment
         android:id="@+id/keyboardFragment"
@@ -30,11 +126,10 @@
         <action
             android:id="@+id/action_keyboardFragment_to_addToCategoryPickerFragment"
             app:destination="@id/addToCategoryPickerFragment" />
+        <action
+            android:id="@+id/action_keyboardFragment_to_settingsFragment"
+            app:destination="@id/settingsFragment" />
     </fragment>
-    <activity
-        android:id="@+id/settingsActivity"
-        android:name="com.willowtree.vocable.settings.SettingsActivity"
-        android:label="SettingsActivity" />
     <fragment
         android:id="@+id/addToCategoryPickerFragment"
         android:name="com.willowtree.vocable.presets.AddToCategoryPickerFragment"
